Replace int with Py_ssize_t anywhere dealing with memory offsets
Enables matches where start= and end= exceed 2**31 (2 GiB), fixes crash
on OverflowError

+++ src/re2.pyx
@@ -311,7 +311,7 @@ cdef inline unicode cpp_to_unicode(cpp_string input):
             input.data(), input.length(), 'strict')
 
 
-cdef inline unicode char_to_unicode(const char * input, int length):
+cdef inline unicode char_to_unicode(const char * input, Py_ssize_t length):
     """Convert a C string to a unicode string."""
     return cpython.unicode.PyUnicode_DecodeUTF8(input, length, 'strict')
 
@@ -368,13 +368,13 @@ cdef inline void release_cstring(Py_buffer *buf):
         PyBuffer_Release(buf)
 
 
-cdef utf8indices(char * cstring, int size, int *pos, int *endpos):
+cdef utf8indices(char * cstring, Py_ssize_t size, Py_ssize_t *pos, Py_ssize_t *endpos):
     """Convert unicode indices ``pos`` and ``endpos`` to UTF-8 indices.
 
     If the indices are out of range, leave them unchanged."""
     cdef unsigned char * data = <unsigned char *>cstring
-    cdef int newpos = pos[0], newendpos = -1
-    cdef int cpos = 0, upos = 0
+    cdef Py_ssize_t newpos = pos[0], newendpos = -1
+    cdef Py_ssize_t cpos = 0, upos = 0
     while cpos < size:
         if data[cpos] < 0x80:
             cpos += 1
@@ -405,11 +405,11 @@ cdef utf8indices(char * cstring, int size, int *pos, int *endpos):
     endpos[0] = newendpos
 
 
-cdef void unicodeindices(map[int, int] &positions,
-        char * cstring, int size, int * cpos, int * upos):
+cdef void unicodeindices(map[Py_ssize_t, Py_ssize_t] &positions,
+        char * cstring, Py_ssize_t size, Py_ssize_t * cpos, Py_ssize_t * upos):
     """Convert UTF-8 byte indices to unicode indices."""
     cdef unsigned char * s = <unsigned char *>cstring
-    cdef map[int, int].iterator it = positions.begin()
+    cdef map[Py_ssize_t, Py_ssize_t].iterator it = positions.begin()
 
     if dereference(it).first == -1:
         dereference(it).second = -1
